almost done

+++ internal/service/history/create.go
@@ -5,7 +5,7 @@ import (
 	modelserv "go-caro/internal/service/history/model"
 )
 
-func (s *service) Create(ctx context.Context, post *modelserv.PostHistory) (uint64, error) {
+func (s *service) Create(ctx context.Context, post *modelserv.PostHistory) (int, error) {
 	//TODO: add transactions
 	id, err := s.historyRepo.Create(ctx, post)
 	if err != nil {
