+++ internal/service/history/delete_by_id.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (s *service) DeleteByID(ctx context.Context, id uint64) error {
+func (s *service) DeleteByID(ctx context.Context, id int) error {
 	//TODO: add transactions
 	err := s.historyRepo.DeleteByID(ctx, id)
 	if err != nil {
